+++ src/core/BuildSystem.ts
@@ -4,7 +4,7 @@ import { ModelRouter } from './ModelRouter';
 import { DebugEngine } from './DebugEngine';
 import { StorageManager } from '../services/StorageManager';
 import { Logger } from '../utils/Logger';
-import { DebugLog } from '../utils/DebugLog';
+import { UltraDevLog as DebugLog } from '../utils/UltraDevLog';
 import { AppArchitect } from './AppArchitect';
 import { ProjectGenerator } from './ProjectGenerator';
 import { MavenResolver } from './MavenResolver';